+++ docsrc/conf.py
@@ -99,6 +99,8 @@ html_context = {
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
 
+html_favicon = '../resources/icons/porymap-icon-2.ico'
+
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
 #