+++ src/log.cpp
@@ -86,3 +86,14 @@ QString getLogPath() {
 QString getMostRecentError() {
     return mostRecentError;
 }
+
+bool cleanupLargeLog() {
+    QFile logFile(getLogPath());
+    if (logFile.size() < 20000000)
+        return false;
+
+    bool removed = logFile.remove();
+    if (removed)
+        logWarn(QString("Previous log file %1 was cleared due to being over 20MB in size.").arg(getLogPath()));
+    return removed;
+}